+++ receiver/hostmetricsreceiver/internal/scraper/memoryscraper/memory_scraper.go
@@ -18,7 +18,9 @@ import (
 	"context"
 	"time"
 
+	"github.com/shirou/gopsutil/v3/host"
 	"github.com/shirou/gopsutil/v3/mem"
+	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/model/pdata"
 	"go.opentelemetry.io/collector/receiver/scrapererror"
 
@@ -30,17 +32,28 @@ const metricsLen = 1
 // scraper for Memory Metrics
 type scraper struct {
 	config *Config
+	mb     *metadata.MetricsBuilder
 
 	// for mocking gopsutil mem.VirtualMemory
 	virtualMemory func() (*mem.VirtualMemoryStat, error)
+	bootTime      func() (uint64, error)
 }
 
 // newMemoryScraper creates a Memory Scraper
 func newMemoryScraper(_ context.Context, cfg *Config) *scraper {
-	return &scraper{config: cfg, virtualMemory: mem.VirtualMemory}
+	return &scraper{config: cfg, virtualMemory: mem.VirtualMemory, bootTime: host.BootTime}
 }
 
-func (s *scraper) Scrape(_ context.Context) (pdata.Metrics, error) {
+func (s *scraper) start(context.Context, component.Host) error {
+	bootTime, err := s.bootTime()
+	if err != nil {
+		return err
+	}
+	s.mb = metadata.NewMetricsBuilder(s.config.Metrics, metadata.WithStartTime(pdata.Timestamp(bootTime*1e9)))
+	return nil
+}
+
+func (s *scraper) scrape(_ context.Context) (pdata.Metrics, error) {
 	md := pdata.NewMetrics()
 	metrics := md.ResourceMetrics().AppendEmpty().InstrumentationLibraryMetrics().AppendEmpty().Metrics()
 
@@ -50,21 +63,7 @@ func (s *scraper) Scrape(_ context.Context) (pdata.Metrics, error) {
 		return md, scrapererror.NewPartialScrapeError(err, metricsLen)
 	}
 
-	metrics.EnsureCapacity(metricsLen)
-	initializeMemoryUsageMetric(metrics.AppendEmpty(), now, memInfo)
+	s.recordMemoryUsageDataPoints(now, memInfo)
+	s.mb.Emit(metrics)
 	return md, nil
-}
-
-func initializeMemoryUsageMetric(metric pdata.Metric, now pdata.Timestamp, memInfo *mem.VirtualMemoryStat) {
-	metadata.Metrics.SystemMemoryUsage.Init(metric)
-
-	idps := metric.Sum().DataPoints()
-	idps.EnsureCapacity(memStatesLen)
-	appendMemoryUsageStateDataPoints(idps, now, memInfo)
-}
-
-func initializeMemoryUsageDataPoint(dataPoint pdata.NumberDataPoint, now pdata.Timestamp, stateLabel string, value int64) {
-	dataPoint.Attributes().InsertString(metadata.Attributes.State, stateLabel)
-	dataPoint.SetTimestamp(now)
-	dataPoint.SetIntVal(value)
 }